fix(core): use console to show error if ErrorHandler not present
The errors used to be swallowed if no `Injector` or `ErrorHandler` were present. The new implementation falls back to `console` or re-throws if no `console` is present.

Using classical bootstrap we are guaranteed to have `Injector`. This can only happen if we bootstrap using `renderComponent` which is not public yet and so it it is used in tests only.